Move ClientServiceWithIndexedDb to another file
